Finished implementing finite state machine
I finished the implementation of the finite state machine and found a bug causing issues with both AIs. In the moveSnake function when the snake moved up the new empty space created when the snake moved was being stored in the snaketrix as 0 and not 2 which all other empty spaces were stored as so was causing both algorithms to avoid.